requirements.txt
openai>=1.30.0


langchain>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1

chromadb>=0.5.5

beautifulsoup4>=4.12.2
lxml>=4.9.4
requests>=2.31.0

gradio==4.44.0
gradio-client==1.3.0
typer==0.12.5

python-dotenv>=1.0.1
